[MCTPD] Discover MCTP capable SMBus device
This patch implements code to discover SMBus devices that support the
Management Component Transport Protocol (MCTP).

This runs a scan on the preconfigured i2c bus and the associated mux
buses. If there are MCTP capable devices (which respond to
getMCTPVersion command), they are registered and the EID is represented on
D-Bus, and the necessary information about the MCTP endpoint
like supported MCTP message types are populated as D-Bus properties.

Also, added a fix in test-mctpd to link i2c library.

Tested:
Started mctpd instance on a preconfigured port and noticed that EID was
assigned and represented on D-Bus for the SMBus devices which support MCTP.

1. Cards behind the mux were discovered successfully.
2. Observed that the flow for registering MCTP endpoint was triggered
successfully for the discovered device.
(Get EID, Set EID, Get UUID, Get Supported Message Types)
3. Started mctpd instance on a port where there were no SMBus devices
with MCTP support, observed that no device was enumerated on D-Bus.

Standalone builds for both mctpd and test-mctpd passed.

Change-Id: I2095e5f16a2830ceb17e1df35db77ee8a61cee97
Signed-off-by: Sumanth Bhat <sumanth.bhat@intel.com>