Remove and replace preparePrivateDataResp()
This change removes preparePrivateDataResp() method and replaces it
with a virtual method named isReceivedPrivateDataCorrect(). This new
method is called early in dispatcher.

Tested:

Performed tests were similar to those performed for discovery flow and
getting routing table. No changes in behavior were observed.

The new method is verified to be called when we enter
MctpBinding::handleCtrlReq().

Change-Id: I4ed53c8476153d3c3eede07ae592a292d28eae06
Signed-off-by: Michal Gleba <michal.gleba@intel.com>