[mctpd]Assign UUID based EID for hotplug devices
EID assignment is done based on UUID in order
-->To avoid depletion of EIDs in the pool,
-->To avoid an endpoint getting registered more than once unnecessarily
-->Also to ensure that the endpoint which is unregistered due to reset
is regsitered back with the right eid from the pool.

Also added support to register hotplug devices.

uUidTable is introduced here, which holds the entry of eid and uuid of
each endpoint registered.

uUidTable and deviceTable's are taken care of clearing the eid entry
when the endpoint is unregistered(If the endpoint undergoes reset or it
is removed)

EID returing to the pool for further assignment is done in the
required scenarios.

The eidPool maintained by the busOwner is now implemented to be in FIFO
list of reclaimed EIDs, so that the older EID from the pool is picked
for registered the endpoint.

Device removal check: Treclaim interval(5 Sec) as mentioned in spec
is not added, as 60 Sec wait interval of scan loop is good enough
to check whether the device is removed.

Removed device is handled by unregistering the EID assigned
and returning its EID to the eidPool.

Test:
Endpoint registration is done based on the UUID assigned.
The device which is already registered(EID=9) is ignored to be registered in
every minute scan loop.
The device which is reset due to Firmware update, is unregistered and
registered back with the new EID from the pool.
EID=9 of unregistered endpoint is released to the pool for next
assignment.
The EID=10 is taken from pool for next assignation of the endpoint.
Verified pldm level device inits/removal is getting triggered correspondingly.

From the test logs:
May 19 08:53:52 intel-obmc mctpd[492]: Checking device 50 to be registered
May 19 08:53:52 intel-obmc mctpd[492]: New Device Registered: EID=14
May 19 08:53:52 intel-obmc mctpd[492]: Checking device 97 to be registered
.
.
.
May 19 08:53:54 intel-obmc mctpd[492]: Checking device 88 to be registered
May 19 08:53:57 intel-obmc mctpd[492]: New Device Registered: EID=9
.
.
.
Sep 18 23:41:15 intel-obmc mctpd[492]: Checking device 50 to be registered
Sep 18 23:41:16 intel-obmc mctpd[492]: Retry timed out, No response
Sep 18 23:41:16 intel-obmc mctpd[492]: Get MCTP Version Support: Unable to get response
Sep 18 23:41:16 intel-obmc mctpd[492]: Get MCTP Control Version failed
Sep 18 23:41:16 intel-obmc mctpd[492]: Unregistering EID = 14
.
.
.
Sep 18 23:41:16 intel-obmc mctpd[492]: Checking device 88 to be registered
Sep 18 23:41:16 intel-obmc mctpd[492]: Endpoint already Registered with EID 9

Change-Id: I11cf996400b75d6225139c0fbbc798ed7ffc6dee
Signed-off-by: Ashwini N S <ashwinix.niduvani.shivashankarappa@intel.com>
Signed-off-by: Arun P. Mohanan <arun.p.m@linux.intel.com>