[MCTPD]Fix off-by-one issue in getVDPCIMessageTypes
Last vendor defined message type/set wasn't getting added to
the supported Vendor Defined Message Types list due to loop
being broken if we encountered "0xFF" (no-more-vendor-defined-message-set).
This provides a fix for the same.

Tested:
Ran it on a device which supports single vendor defined message type.
The type was populated on D-Bus

Change-Id: Ia7d6a967e562bad4b7f0822260d7f659f6727d4e
Signed-off-by: Sumanth Bhat <sumanth.bhat@intel.com>