[MCTP]Update MCTP endpoint interfaces to a map
Interfaces exposed on D-Bus are the final result of device discovery.
Updating the interfaces to use a map which holds EID gives the
advantages to directly validate the endpoints(EID) discovered.
Also there can be endpoints with static eid's whose eid is not part
of our EID pool. This change add initial support to cache the
interfaces along with associated EID.

Adding/removing interfaces code is updated to accomodate the change to
maps.
isEIDRegistered() is updated to use the map(endpointInterface) to
identify whether EID is registered or not.

Test:
Endpoint Interfaces exposed to D-Bus as expected.
Interfaces getting deleted on device removal.

Change-Id: Ibf0f3b6bfab3e7a5eadb70a3542095626f2f8049
Signed-off-by: Ashwini N S <ashwinix.niduvani.shivashankarappa@intel.com>