[mctpd] Add TransportBindingType config option
In entity manager, if a configuration node has a property starting
with "Bind", it will try to bind this node to the node to the node
this property points to. In current mctpd implementation, we use a
property "BindingType" to tell the MCTP transport binding, it will
confuse the entity manager and make it give out errors. This patch
fixes this issue by adding a new option "TransportBindingType" to
replace it.

The old "BindingType" is deprecated, but it's kept for compatibility
concerns and will be removed later.

Tested:
Renamed BindingType to TransportBindingType in entity manager json,
and there is no behavior differences.

Change-Id: I4eaff4f25ae16ea158e0344e545974fbd76b0b39
Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>