Fix EID rediscovery flow (#32)
Ideally, when a device(which supports UUID) loose its EID(as part of
unplanned device reset), mctpd is supposed to assign back same EID
back from UUID table. This was not happening due to a bug in the flow.

Update the code to take EID from UUID table in the registration flow.
Modified few functions as part of cleanup.

During testing it was observed 3 sec delay in not good enough for
the MCTP device to power off and disappear from SMBus. Thus increased
rediscovery delay to 5sec.

Tested:
Triggered device re-registration and verified the device is getting
re-registered as expected.

Signed-off-by: Arun P. Mohanan <arun.p.m@linux.intel.com>