SMBusBinding: Fix muxMonitor inotify event name handling (#47)
In kernel inotify, the inotify_event.name is padded to a mutiple of
sizeof(inotify_event) with '\0' and len is the size of char[] name,
not the actual size of name. So constructing the name string with
std::string(name, len) constructs a string with all the '\0's, which
is not equal to the filename. Use std::string(name) instead so that
the string does not contain these '\0's.

Tested:
Verified i2c bus changes are handled properly when loading/unloading
i2c mux driver.

Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>