PCIeBinding: Handle endpoint registration error
When processing routing table changes for PCIe binding, mctpd does not
check the return value of registerEndpoint. But the registration will
fail in some cases, like the endpoint fails to respond the control
message within the time limit (120 ms). This commit handles such errors
by removing the endpoints failed to register in the routing table, so
mctpd will try reregistering it in next routing table update.

Tested:
Verified the device is not added to routing table when registerEndpoint
returns std::nullopt, and mctpd tries to reregister it in next update.

Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>